refactor(tape): drop seams the projection never used
Removing each piece the branch added and asking what breaks left these
with no answer. TapeTranscriptReader had one remaining reader, the
projection port that extended it; the port now declares getMessages
itself. The row-shape types in messageContent.ts and the search-text
extractor in the applier were exported without an importer. The
divergence guard only runs for live appends, so its source parameter and
log field were constant. commitRetractions took the ids next to the
records they came from. The range-read SQL builder took an options bag
for one boolean.

The usage-stats writer moves into the transcript as a private method:
after the previous fix only the transcript's assistant terminal writes
call it, so it no longer belongs next to the projection applier.
